feat(sdk): WorkspaceOps interface on the runtime seam
Separate optional interface (RFC runtime-evolution.md §4.2). A runtime may
implement it; the desktop host always can. OpenCodeClient implements it
next; kimi will not (validates the 'separate, not merged' design).

The interface is in packages/sdk/src/workspace.ts and re-exported from the
SDK barrel. Its types come from @ai4s/shared:

- DirEntry and ArtifactFile were previously defined locally in
  apps/desktop/src/lib/artifactFile.ts and NOT exported from @ai4s/shared
  (only FileRoot was). The SDK cannot import from apps/desktop (wrong
  dependency direction), so the two stable domain types are promoted to
  packages/shared/src/index.ts (exact existing definitions — no invention)
  and artifactFile.ts now re-exports them from shared. All existing callsites
  that import these from @/lib/artifactFile keep working unchanged.

TDD: apps/desktop/src/test/workspace.test.ts asserts the four-method contract.
The test is a type-level contract (the value assertions are on a plain
object), so it cannot fail at runtime when WorkspaceOps is absent — the
guarantee is enforced by tsc, which now compiles cleanly across desktop,
sdk, and shared.
